+++ main/src/main/java/net/oneandone/stool/util/Vhost.java
@@ -29,16 +29,14 @@ import java.util.Objects;
 public class Vhost {
     private static final char SEP = ' ';
 
-    // parses   <even> <name> <stage> <id> [<docroot>]
-    // where name is the application name and stage is the stage name.
+    // parses   <even> <name> <id> [<docroot>]
+    // where name is the application name.
     public static Vhost forLine(World world, String line) {
         int afterEven;
         int afterName;
-        int afterStage;
         int afterId;
         int even;
         String name;
-        String stage;
         String id;
         FileNode docroot;
 
@@ -54,30 +52,21 @@ public class Vhost {
         }
         name = line.substring(afterEven + 1, afterName);
 
-        afterStage = line.indexOf(SEP, afterName + 1);
-        if (afterStage == -1) {
-            throw new IllegalArgumentException("invalid vhost line: " + line);
-        }
-        stage = line.substring(afterName + 1, afterStage);
-
-        afterId = line.indexOf(SEP, afterStage + 1);
+        afterId = line.indexOf(SEP, afterName + 1);
         if (afterId == -1) {
-            id = line.substring(afterStage + 1);
+            id = line.substring(afterName + 1);
             docroot = null;
         } else {
-            id = line.substring(afterStage + 1, afterId);
+            id = line.substring(afterName + 1, afterId);
             docroot = world.file(line.substring(afterId + 1));
         }
-        return new Vhost(even, name, stage, id, docroot);
+        return new Vhost(even, name, id, docroot);
     }
 
     public final int even;
 
-    /** name of the application */
+    /** name of vhost, i.e. the application */
     public final String name;
-
-    /** TODO: redundant, dump ... */
-    public final String stage;
 
     /** stage id */
     public final String id;
@@ -85,12 +74,9 @@ public class Vhost {
     /** null for ports that have no domain */
     public final FileNode docroot;
 
-    public Vhost(int even, String name, String stage, String id, FileNode docroot) {
+    public Vhost(int even, String name, String id, FileNode docroot) {
         if (name.indexOf(SEP) != -1) {
             throw new IllegalArgumentException(name);
-        }
-        if (stage.indexOf(SEP) != -1) {
-            throw new IllegalArgumentException(stage);
         }
         if (id.indexOf('.') == -1) {
             throw new IllegalArgumentException(id);
@@ -98,7 +84,6 @@ public class Vhost {
         this.even = even;
         this.name = name;
         this.id = id;
-        this.stage = stage;
         this.docroot = docroot;
     }
 
@@ -118,25 +103,25 @@ public class Vhost {
         return even + 1;
     }
 
-    public String httpUrl(boolean vhosts, String hostname) {
-        return "http://" + fqdnHttpPort(vhosts, hostname);
+    public String httpUrl(boolean vhosts, String stageName, String hostname) {
+        return "http://" + fqdnHttpPort(vhosts, stageName, hostname);
     }
 
-    public String fqdnHttpPort(boolean vhosts, String hostname) {
-        return fqdn(vhosts, hostname) + ":" + httpPort();
+    public String fqdnHttpPort(boolean vhosts, String stageName, String hostname) {
+        return fqdn(vhosts, stageName, hostname) + ":" + httpPort();
     }
 
-    public String httpsUrl(boolean vhosts, String hostname) {
-        return "https://" + fqdnHttpsPort(vhosts, hostname);
+    public String httpsUrl(boolean vhosts, String stageName, String hostname) {
+        return "https://" + fqdnHttpsPort(vhosts, stageName, hostname);
     }
 
-    public String fqdnHttpsPort(boolean vhosts, String hostname) {
-        return fqdn(vhosts, hostname) + ":" + httpsPort();
+    public String fqdnHttpsPort(boolean vhosts, String stageName, String hostname) {
+        return fqdn(vhosts, stageName, hostname) + ":" + httpsPort();
     }
 
-    public String fqdn(boolean vhosts, String hostname) {
+    public String fqdn(boolean vhosts, String stageName, String hostname) {
         if (vhosts) {
-            return name + "." + stage + "." + hostname;
+            return name + "." + stageName + "." + hostname;
         } else {
             return hostname;
         }
@@ -146,7 +131,7 @@ public class Vhost {
         // CAUTION: just
         //    even + SEP
         // is an integer addition!
-        return Integer.toString(even) + SEP + name + SEP + stage + SEP + id + (docroot == null ? "" : Character.toString(SEP) + docroot.getAbsolute());
+        return Integer.toString(even) + SEP + name + SEP + id + (docroot == null ? "" : Character.toString(SEP) + docroot.getAbsolute());
     }
 
     public String toString() {
@@ -158,15 +143,15 @@ public class Vhost {
         if (Objects.equals(this.docroot, newDocroot) && (newEven == null || newEven == even)) {
             return null;
         }
-        return new Vhost(newEven == null ? even : newEven, name, stage, id, newDocroot);
+        return new Vhost(newEven == null ? even : newEven, name, id, newDocroot);
     }
 
-    public String context(String hostname, String url) {
+    public String context(String stageName, String hostname, String url) {
         String result;
         String context;
 
         result = null;
-        for (String str : doMap(hostname, url).values()) {
+        for (String str : doMap(stageName, hostname, url).values()) {
             context = getContext(str);
             if (result == null) {
                 result = context;
@@ -183,11 +168,11 @@ public class Vhost {
         return result;
     }
 
-    public Map<String, String> urlMap(String hostname, String url) {
+    public Map<String, String> urlMap(String stageName, String hostname, String url) {
         Map<String, String> result;
 
         result = new LinkedHashMap<>();
-        for (Map.Entry<String, String> entry : doMap(hostname, url).entrySet()) {
+        for (Map.Entry<String, String> entry : doMap(stageName, hostname, url).entrySet()) {
             result.put(entry.getKey(), hideContext(entry.getValue()));
         }
         return result;
@@ -231,7 +216,7 @@ public class Vhost {
         return context == -1 ? "" : url.substring(afterHost, context);
     }
 
-    private Map<String, String> doMap(String hostname, String url) {
+    private Map<String, String> doMap(String stageName, String hostname, String url) {
         Map<String, String> result;
         Map<Character, String> map;
         List<String> all;
@@ -242,7 +227,7 @@ public class Vhost {
         map = new HashMap<>();
         map.put('h', hostname);
         map.put('a', name);
-        map.put('s', stage);
+        map.put('s', stageName);
         map.put('p', "%p");
         all = Url.parse(url).sustitute(map).map();
         http = new ArrayList<>();