+++ CHANGELOG.md
@@ -49,7 +49,8 @@
       org.apache.catalina.core.ContainerBase.[Catalina].handlers = 1catalina.org.apache.juli.FileHandler
 
   to Tomcat logging.properties
-* dumped 'other' field
+* dumped 'other' status field
+* dumped redundant 'stage' entry in ports file
 
 
 ### 3.4.10 (2017-10-05)